Refactoring form into its own component
